WIIS-9176

+++ src/Controller/ZoneController.php
@@ -8,13 +8,7 @@ use App\Entity\Action;
 use App\Entity\Emplacement;
 use App\Entity\Menu;
 use App\Entity\PurchaseRequestScheduleRule;
-use App\Entity\Urgence;
 use App\Entity\Zone;
-use App\Service\CSVExportService;
-use App\Service\SpecificService;
-use App\Service\TranslationService;
-use App\Service\UrgenceService;
-use App\Service\UserService;
 use App\Service\ZoneService;
 use App\Exceptions\FormException;
 use Doctrine\ORM\EntityManagerInterface;
@@ -48,7 +42,7 @@ class ZoneController extends AbstractController
             ->setName($request->request->get("name"))
             ->setDescription($request->request->get("description"))
             ->setInventoryIndicator($request->request->get("inventoryIndicator") ?? null)
-            ->setActive($request->request->get("active"));
+            ->setActive($request->request->getBoolean("active"));
 
         $manager->persist($zone);
         $manager->flush();
@@ -98,7 +92,7 @@ class ZoneController extends AbstractController
                     . ( $purchaseRequestScheduleRuleRepository->isZoneInPurchaseRequestScheduleRule($zone) ? ' une planification de demande d’achat, ' : '');
 
                 if ($issue) {
-                    throw new FormException("La zone ou ses emplacements sont contenus dans".$issue."vous ne pouvez donc pas la désactiver");
+                    throw new FormException("La zone ou ses emplacements sont contenus dans" . $issue . "vous ne pouvez donc pas la désactiver");
                 }
             }
 
@@ -120,39 +114,19 @@ class ZoneController extends AbstractController
         throw new NotFoundHttpException();
     }
 
-    #[Route("/api-supprimer", name: "zone_delete_api", options: ["expose" => true], methods: "GET|POST", condition: "request.isXmlHttpRequest()")]
+    #[Route("/{zone}/delete", name: "zone_delete", options: ["expose" => true], methods: "DELETE", condition: "request.isXmlHttpRequest()")]
     #[HasPermission([Menu::REFERENTIEL, Action::DELETE], mode: HasPermission::IN_JSON)]
-    public function deleteApi(Request $request, EntityManagerInterface $manager): Response {
-        if ($request->isXmlHttpRequest() && $data = json_decode($request->getContent(), true)) {
-            $zoneRepository = $manager->getRepository(Zone::class);
-            $zone = $zoneRepository->find($data["id"]);
-
-            if ($zone->getLocations()->isEmpty()){
-                $delete = true;
-                $html = $this->renderView('zone/delete_content.html.twig');
-
-            } else {
-                $delete = false;
-                $html = $this->renderView('zone/delete_content_wrong.html.twig');
-            }
-
-            return new JsonResponse(['delete' => $delete, 'html' => $html]);
-        }
-
-        throw new BadRequestHttpException();
-    }
-
-    #[Route("/supprimer", name: "zone_delete", options: ["expose" => true], methods: "POST", condition: "request.isXmlHttpRequest()")]
-    #[HasPermission([Menu::REFERENTIEL, Action::DELETE], mode: HasPermission::IN_JSON)]
-    public function delete(Request $request,
-                           EntityManagerInterface $entityManager): Response
-    {
-        $data = $request->request->all();
-        $zoneRepository = $entityManager->getRepository(Zone::class);
-        $zone = $zoneRepository->find($data['id']);
+    public function delete(Zone                   $zone,
+                           EntityManagerInterface $entityManager): Response {
+        $purchaseRequestScheduleRuleRepository = $entityManager->getRepository(PurchaseRequestScheduleRule::class);
 
         if (!$zone->getLocations()->isEmpty()){
-            throw new FormException("Vous ne pouvez pas supprimer cette zone car elle est lié à un ou plusieur(s) emplacements.");
+            throw new FormException("Vous ne pouvez pas supprimer cette zone car elle est liée à un ou plusieurs emplacements. Vous pouvez la rendre inactive en modifiant la zone.");
+        }
+
+        $purchaseRequestCounter = $purchaseRequestScheduleRuleRepository->count(['zones' => $zone]);
+        if ($purchaseRequestCounter > 0){
+            throw new FormException("Vous ne pouvez pas supprimer cette zone car elle est lié à une ou plusieurs planifications de demandes d'achat. Vous pouvez la rendre inactive en modifiant la zone.");
         }
 
         $entityManager->remove($zone);
